requirements.txt
openpyxl>=3.1.0
python-docx>=0.8.11
Pillow>=10.0.0
pywin32>=306

